+++ ReleaseNotes.md
@@ -2,4 +2,4 @@ Message Board Release Notes
 
 Friday, April 17, 2020
 
-* Fixed: Unable to save a new thread.*
+* Fixed: Unable to save a new message to a thread.*